requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
wandb>=0.16.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
rasterio>=1.3.0
tqdm>=4.65.0
Pillow>=10.0.0
pyyaml>=6.0
loguru>=0.7.0
gdown>=4.7.0
google-cloud-storage>=2.10.0
